requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
google-genai==1.0.0
